use double as return type for moving product since int return type overflows too quickly

+++ tslib/vector.summary/prod.hpp
@@ -35,7 +35,7 @@ namespace tslib {
   template<>
   class prodTraits<int> {
   public:
-    typedef int ReturnType;
+    typedef double ReturnType;
   };
